Allow folders for the train and validation paths.
This changes in the `taming/data/custom.py` file allow the use of folders for the training and validation paths while been compatible with how it was used before. This should make it easier to train on custom data as you no longer need to create a `train.txt` and `val.txt` file manually which sometimes can be a time-consuming and tedious task, now you can just create two folders, one for the train and one for the validation containing the images inside and the script will discover the files and create the list of paths by itself, this can also allow you to add more images to those folders later without having to manually recreate the train and validation files, not sure about the use for that but the option is there.